requirements.txt
certifi==2018.4.16
chardet==3.0.4
confusable-homoglyphs==3.1.1
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.36.0
django-cors-headers==2.3.0
django-sendgrid-v5==0.6.893
djangorestframework==3.8.2
future==0.16.0
gunicorn==19.6.0
idna==2.7
oauthlib==2.1.0
Pillow==5.0.0
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
sendgrid==5.4.1
urllib3==1.23
whitenoise==3.3.1